fix: auto height issue on dialogs
